renamed DomainField to ApiUrl
